fix: Float/Float16 eⁿ via Double round-trip for Android portability
Android's NDK math.h exposes only exp(Double) — no Float overload — so exp(Float)
failed to compile. Float(exp(Double(n))) is identical and portable on every platform.

+++ Sources/RealNumber/Float16+Real.swift
@@ -10,7 +10,8 @@ import Foundation
         }
 
         public static func eⁿ(_ n: Self) -> Self {
-            Float16(exp(Float(n)))
+            // Round-trip through Double (Android's NDK exposes only exp(Double)).
+            Float16(exp(Double(n)))
         }
     }
 #endif
